+++ src/lib/logseq.ts
@@ -65,7 +65,7 @@ export async function saveDalleImage(imageURL: string): Promise<string> {
   return imageFileName;
 }
 
-export async function getAudioFile(content: string): Promise<File| null> {
+export async function getAudioFile(content: string): Promise<File | null> {
   //extract file path in () from markdown link like ![my file](assets/2023-03-17-13-24-36.m4a)
   //supported formats are mp3, mp4, mpeg, mpga, m4a, wav, and webm
   const path = (await logseq.App.getCurrentGraph())?.path;
@@ -79,7 +79,7 @@ export async function getAudioFile(content: string): Promise<File| null> {
   if (!extension) {
     return null;
   }
-//remove ../ from path
+  //remove ../ from path
   const filepath = match[1].replace("../", "");
   // get filename from path by removing assets/ from path
   const filename = filepath.replace("assets/", "");
@@ -89,3 +89,4 @@ export async function getAudioFile(content: string): Promise<File| null> {
   const file = new File([audioBlob], filename, { type: `audio/${extension}` });
   return file;
 }
+
