+++ src/main.tsx
@@ -1,15 +1,14 @@
 import "./ui/style.css";
 import "@logseq/libs";
-import { openAI, whisper } from "./lib/openai";
+import { openAI } from "./lib/openai";
 import React, { useState } from "react";
 import ReactDOM from "react-dom/client";
 import { Command, LogseqAI } from "./ui/LogseqAI";
 import { loadUserCommands, loadBuiltInCommands } from "./lib/prompts";
 import { getOpenaiSettings, settingsSchema } from "./lib/settings";
-import { runDalleBlock, runGptBlock, runGptPage } from "./lib/rawCommands";
+import { runDalleBlock, runGptBlock, runGptPage, runWhisper } from "./lib/rawCommands";
 import { BlockEntity } from "@logseq/libs/dist/LSPlugin.user";
 import { useImmer } from 'use-immer';
-import { getAudioFile } from "./lib/logseq";
 
 logseq.useSettingsSchema(settingsSchema);
 
@@ -160,14 +159,14 @@ const LogseqApp = () => {
         openUI();
       }
     });
-    logseq.Editor.registerSlashCommand("whisper", runWhisper);
-
     logseq.Editor.registerSlashCommand("gpt-page", runGptPage);
     logseq.Editor.registerBlockContextMenuItem("gpt-page", runGptPage);
     logseq.Editor.registerSlashCommand("gpt-block", runGptBlock);
     logseq.Editor.registerBlockContextMenuItem("gpt-block", runGptBlock);
     logseq.Editor.registerSlashCommand("dalle", runDalleBlock);
     logseq.Editor.registerBlockContextMenuItem("dalle", runDalleBlock);
+    logseq.Editor.registerSlashCommand("whisper", runWhisper);
+    logseq.Editor.registerBlockContextMenuItem("whisper", runWhisper);
 
     if (logseq.settings!["shortcutBlock"]) {
       logseq.App.registerCommandShortcut(
@@ -176,19 +175,7 @@ const LogseqApp = () => {
       );
     }
   }, []);
-  async function runWhisper() {
-    const currentBlock = await logseq.Editor.getCurrentBlock();
-    if (currentBlock) {
-      const audioFile = await getAudioFile(currentBlock.content);
-      if (audioFile) {
-        const openAISettings = getOpenaiSettings();
-        const transcribe = await whisper(audioFile, openAISettings);
-        if (transcribe) {
-          await logseq.Editor.insertBlock(currentBlock.uuid, transcribe);
-        }
-      }
-  }
-}
+
 
   const allCommands = [...builtInCommands, ...userCommands];
